# day02-tlias-web-management/Web开发-2024.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Web开发-2024",
    "description": "新版本的Web开发的接口文档",
    "version": "1.0.0"
  },
...
  "paths": {
    "/clazzs": {
      "get": {
        "summary": "分页查询",
        "x-apifox-folder": "班级管理",
        "x-apifox-status": "released",
        "deprecated": false,
        "description": "班级信息的分页条件查询",
        "operationId": "pageUsingGET",
        "tags": [
          "班级管理"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "班级名称",
            "required": false,
            "example": "java就业100期",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "begin",
            "in": "query",
            "description": "结课时间的开始时间",
            "required": false,
            "example": "2020-01-01",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "end",
            "in": "query",
            "description": "结课时间的结束时间",
            "required": false,
            "example": "2023-01-01",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "页码",
            "required": false,
            "example": "1",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "每页展示记录数",
            "required": false,
            "example": "10",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "x-apifox-refs": {},
                  "x-apifox-orders": [
                    "code",
                    "msg",
                    "data"
                  ],
                  "properties": {
                    "code": {
                      "type": "integer",
                      "title": "状态码",
                      "description": "1: 成功 , 0: 失败",
                      "format": "int32",
                      "additionalProperties": false
                    },
                    "msg": {
                      "type": "string",
                      "title": "错误提示信息",
                      "additionalProperties": false,
                      "nullable": true
                    },
                    "data": {
                      "title": "响应数据",
                      "type": "object",
                      "x-apifox-refs": {},
                      "x-apifox-orders": [
                        "total",
                        "rows"
                      ],
                      "properties": {
                        "total": {
                          "type": "integer",
                          "title": "总记录数",
                          "description": "总记录数",
                          "additionalProperties": false
                        },
                        "rows": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "integer",
                                "format": "int32",
                                "title": "ID",
                                "description": "ID",
                                "additionalProperties": false
                              },
                              "name": {
                                "type": "string",
                                "title": "班级名称",
                                "description": "班级",
                                "additionalProperties": false
                              },
                              "room": {
                                "type": "string",
                                "title": "班级教室",
                                "description": "班级教室",
                                "additionalProperties": false
                              },
                              "beginDate": {
                                "type": "string",
                                "format": "date",
                                "title": "开课时间",
                                "description": "开课时间",
                                "additionalProperties": false
                              },
                              "endDate": {
                                "type": "string",
                                "format": "date",
                                "title": "结课时间",
                                "description": "结课时间",
                                "additionalProperties": false
                              },
                              "masterId": {
                                "type": "integer",
                                "format": "int32",
                                "title": "班主任id",
                                "description": "班主任id, 关联员工表id",
                                "additionalProperties": false
                              },
                              "masterName": {
                                "type": "string",
                                "title": "班主任姓名",
                                "description": "班主任姓名"
                              },
                              "subject": {
                                "type": "integer",
                                "format": "int32",
                                "title": "学科",
                                "description": "1:java, 2:前端, 3:大数据, 4:Python, 5:Go, 6:嵌入式",
                                "additionalProperties": false
                              },
                              "createTime": {
                                "type": "string",
                                "format": "date-time",
                                "title": "创建时间",
                                "description": "创建时间",
                                "additionalProperties": false
                              },
                              "updateTime": {
                                "type": "string",
                                "format": "date-time",
                                "title": "修改时间",
                                "description": "修改时间",
                                "additionalProperties": false
                              }
                            },
                            "x-apifox-refs": {},
                            "x-apifox-orders": [
                              "id",
                              "name",
                              "room",
                              "beginDate",
                              "endDate",
                              "masterId",
                              "masterName",
                              "subject",
                              "createTime",
                              "updateTime"
                            ],
                            "x-apifox-ignore-properties": []
                          },
                          "title": "结果列表",
                          "description": "结果列表"
                        }
                      },
                      "required": [
                        "rows",
                        "total"
                      ],
                      "x-apifox-ignore-properties": [],
                      "nullable": true
                    }
                  },
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3161925/apis/api-103410783-run"
      },
      "post": {
        "summary": "添加班级",
        "x-apifox-folder": "班级管理",
        "x-apifox-status": "released",
        "deprecated": false,
        "description": "添加班级数据",
        "operationId": "saveUsingPOST",
        "tags": [
          "班级管理"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "title": "班级名称",
                    "description": "班级"
                  },
                  "room": {
                    "type": "string",
                    "title": "班级教室",
                    "description": "班级教室"
                  },
                  "beginDate": {
                    "type": "string",
                    "format": "date",
                    "title": "开课时间",
                    "description": "开课时间"
                  },
                  "endDate": {
                    "type": "string",
                    "format": "date",
                    "title": "结课时间",
                    "description": "结课时间"
                  },
                  "masterId": {
                    "type": "integer",
                    "format": "int32",
                    "title": "班主任id",
                    "description": "班主任id, 关联员工表id"
                  },
                  "subject": {
                    "type": "integer",
                    "format": "int32",
                    "title": "学科",
                    "description": "1:java, 2:前端, 3:大数据, 4:Python, 5:Go, 6:嵌入式"
                  }
                },
                "x-apifox-refs": {},
                "x-apifox-orders": [
                  "name",
                  "room",
                  "beginDate",
                  "endDate",
                  "masterId",
                  "subject"
                ],
                "x-apifox-ignore-properties": []
              },
              "example": ""
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3161925/apis/api-103410784-run"
      },
      "put": {
        "summary": "修改班级",
        "x-apifox-folder": "班级管理",
        "x-apifox-status": "released",
        "deprecated": false,
        "description": "修改班级数据信息",
        "operationId": "updateUsingPUT",
        "tags": [
          "班级管理"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Clazz"
              },
              "example": ""
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3161925/apis/api-103410785-run"
      }
    },
...
    "/clazzs/{id}": {
      "get": {
        "summary": "根据ID查询",
        "x-apifox-folder": "班级管理",
        "x-apifox-status": "released",
        "deprecated": false,
        "description": "根据ID查询班级信息",
        "operationId": "getInfoUsingGET",
        "tags": [
          "班级管理"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "example": "1",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "x-apifox-refs": {
                    "01H7YTX04J7GX93MGMS37EFYGC": {
                      "$ref": "#/components/schemas/Result"
                    }
                  },
                  "x-apifox-orders": [
                    "01H7YTX04J7GX93MGMS37EFYGC"
                  ],
                  "properties": {
                    "code": {
                      "type": "integer",
                      "title": "状态码",
                      "description": "1: 成功 , 0: 失败",
                      "format": "int32"
                    },
                    "msg": {
                      "type": "string",
                      "title": "错误提示信息",
                      "nullable": true
                    },
                    "data": {
                      "title": "响应数据",
                      "anyOf": [
                        {
                          "$ref": "#/components/schemas/Clazz"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "x-apifox-ignore-properties": [
                    "code",
                    "msg",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3161925/apis/api-103410787-run"
      },
      "delete": {
        "summary": "删除班级",
        "x-apifox-folder": "班级管理",
        "x-apifox-status": "released",
        "deprecated": false,
        "description": "根据ID删除班级信息",
        "operationId": "deleteUsingDELETE",
        "tags": [
          "班级管理"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "example": "1",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          },
          "204": {
            "description": "No Content",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3161925/apis/api-103410788-run"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Clazz": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "title": "ID",
            "description": "ID"
          },
          "name": {
            "type": "string",
            "title": "班级名称",
            "description": "班级"
          },
          "room": {
            "type": "string",
            "title": "班级教室",
            "description": "班级教室"
          },
          "beginDate": {
            "type": "string",
            "format": "date",
            "title": "开课时间",
            "description": "开课时间"
          },
          "endDate": {
            "type": "string",
            "format": "date",
            "title": "结课时间",
            "description": "结课时间"
          },
          "masterId": {
            "type": "integer",
            "format": "int32",
            "title": "班主任id",
            "description": "班主任id, 关联员工表id"
          },
          "subject": {
            "type": "integer",
            "format": "int32",
            "title": "学科",
            "description": "1:java, 2:前端, 3:大数据, 4:Python, 5:Go, 6:嵌入式"
          },
          "createTime": {
            "type": "string",
            "format": "date-time",
            "title": "创建时间",
            "description": "创建时间"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time",
            "title": "修改时间",
            "description": "修改时间"
          }
        },
        "title": "Clazz",
        "x-apifox-orders": [
          "id",
          "name",
          "room",
          "beginDate",
          "endDate",
          "masterId",
          "subject",
          "createTime",
          "updateTime"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": "Schemas"
      },
...
      "Result": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "title": "状态码",
            "description": "1: 成功 , 0: 失败",
            "format": "int32"
          },
          "msg": {
            "type": "string",
            "title": "错误提示信息",
            "nullable": true
          },
          "data": {
            "type": "object",
            "properties": {},
            "title": "响应数据",
            "x-apifox-orders": [],
            "x-apifox-ignore-properties": [],
            "nullable": true
          }
        },
        "title": "Result",
        "x-apifox-orders": [
          "code",
          "msg",
          "data"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": "Schemas"
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "开发环境"
    },
    {
      "url": "http://test-cn.your-api-server.com",
      "description": "测试环境"
    },
    {
      "url": "http://prod-cn.your-api-server.com",
      "description": "正式环境"
    }
  ]
}